Add: Black-box false integration tests - scope & order
(cherry picked from commit 3278339399a5263a8a3078c93fff3f6a4811b8d8)
