+++ minitrace-macro/tests/spans/black-box-false-local-async.rs
@@ -3,18 +3,21 @@ use test_utilities::*;
 
 // Reference:
 // - https://github.com/tikv/minitrace-rust/issues/122
-#[trace("test-span")]
-fn f(a: u32) -> u32 {
+#[trace("a-span")]
+async fn f(a: u32) -> u32 {
     a
 }
 
-fn main() {
+#[tokio::main]
+async fn main() {
     let (root, collector) = minitrace::Span::root("root");
-    let _sg1 = root.set_local_parent();
-    f(1);
+    {
+        let _child_span = root.set_local_parent();
+        f(1).await;
+    }
     drop(root);
-    let records: Vec<minitrace::collector::SpanRecord> =
-        futures::executor::block_on(collector.collect());
+    let records: Vec<minitrace::collector::SpanRecord> = futures::executor::block_on(collector.collect());
+
     let expected = r#"[
     SpanRecord {
         id: 1,
@@ -22,6 +25,14 @@ fn main() {
         begin_unix_time_ns: \d+,
         duration_ns: \d+,
         event: "root",
+        properties: [],
+    },
+    SpanRecord {
+        id: 2,
+        parent_id: 1,
+        begin_unix_time_ns: \d+,
+        duration_ns: \d+,
+        event: "a-span",
         properties: [],
     },
 ]"#;
