+++ minitrace-macro/tests/spans/black-box-false-local-sync.rs
@@ -3,19 +3,20 @@ use test_utilities::*;
 
 // Reference:
 // - https://github.com/tikv/minitrace-rust/issues/122
-#[trace("test-span")]
-async fn f(a: u32) -> u32 {
+#[trace("a-span")]
+fn f(a: u32) -> u32 {
     a
 }
 
-#[tokio::main]
-async fn main() {
+fn main() {
     let (root, collector) = minitrace::Span::root("root");
-    let _child_span = minitrace::Span::enter_with_parent("test-span", &root);
-    f(1).await;
+    {
+        let _sg1 = root.set_local_parent();
+        f(1);
+    }
     drop(root);
-    let records: Vec<minitrace::collector::SpanRecord> = futures::executor::block_on(collector.collect());
-
+    let records: Vec<minitrace::collector::SpanRecord> =
+        futures::executor::block_on(collector.collect());
     let expected = r#"[
     SpanRecord {
         id: 1,
@@ -23,6 +24,14 @@ async fn main() {
         begin_unix_time_ns: \d+,
         duration_ns: \d+,
         event: "root",
+        properties: [],
+    },
+    SpanRecord {
+        id: 2,
+        parent_id: 1,
+        begin_unix_time_ns: \d+,
+        duration_ns: \d+,
+        event: "a-span",
         properties: [],
     },
 ]"#;
